refactor overlaytrigger components
