+++ src/options/pages/blueprints/gridView/GridCard.module.scss
@@ -64,8 +64,6 @@
   text-align: right;
   padding-left: 0.5rem;
   white-space: nowrap;
-  text-overflow: ellipsis;
-  cursor: help;
 }
 
 .description {
