+++ src/options/pages/blueprints/labels/LastUpdatedLabel.module.scss
